+++ plugin.video.arteplussept/resources/lib/addon.py
@@ -4,6 +4,7 @@
 #
 # plugin.video.arteplussept, Kodi add-on to watch videos from http://www.arte.tv/guide/fr/plus7/
 # Copyright (C) 2015  known-as-bmf
+# Copyright (C) 2023  thomas-ernest
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License
@@ -19,8 +20,10 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #
+# https://xbmcswift2.readthedocs.io/en/latest/api.html
+# https://github.com/XBMC-Addons/script.module.xbmcswift2
 from xbmcswift2 import Plugin
-
+from xbmcswift2 import xbmc
 
 # global declarations
 # plugin stuff
@@ -34,6 +37,7 @@ class PluginInformation:
 
 # my imports
 import view
+# import api
 from settings import Settings
 
 settings = Settings(plugin)
@@ -41,12 +45,18 @@ settings = Settings(plugin)
 
 @plugin.route('/', name='index')
 def index():
-    return view.build_categories(settings)
+    # return view.build_categories(plugin, plugin.get_storage('cached_categories', TTL=60), settings)
+    return view.build_home_page(plugin, plugin.get_storage('cached_categories', TTL=60), settings)
 
 
-@plugin.route('/category/<category_code>', name='category')
-def category(category_code):
-    return view.build_category(category_code, settings)
+@plugin.route('/api_category/<category_code>', name='api_category')
+def api_category(category_code):
+    return view.build_api_category(category_code, settings)
+
+
+@plugin.route('/cached_category/<category_code>', name='cached_category')
+def cached_category(category_code):
+    return view.get_cached_category(category_code, plugin.get_storage('cached_categories', TTL=60))
 
 
 # @plugin.route('/creative', name='creative')
@@ -58,6 +68,26 @@ def category(category_code):
 def magazines():
     plugin.set_content('tvshows')
     return plugin.finish(view.build_magazines(settings))
+
+
+@plugin.route('/favorites', name='favorites')
+def favorites():
+    plugin.set_content('tvshows')
+    return plugin.finish(view.build_favorites(plugin, settings))
+
+@plugin.route('/add_favorite/<program_id>', name='add_favorite')
+def add_favorite(program_id):
+    view.add_favorite(plugin, settings.username, settings.password, program_id)
+
+@plugin.route('/remove_favorite/<program_id>', name='remove_favorite')
+def remove_favorite(program_id):
+    view.remove_favorite(plugin, settings.username, settings.password, program_id)
+
+
+@plugin.route('/last_viewed', name='last_viewed')
+def favorites():
+    plugin.set_content('tvshows')
+    return plugin.finish(view.build_last_viewed(plugin, settings))
 
 
 @plugin.route('/newest', name='newest')
@@ -90,17 +120,27 @@ def sub_category_by_title(category_code, sub_category_title):
     return plugin.finish(view.build_sub_category_by_title(category_code, sub_category_title, settings))
 
 
-@plugin.route('/collection/<kind>/<collection_id>', name='collection')
-def collection(kind, collection_id):
+@plugin.route('/collection/<kind>/<program_id>', name='collection')
+def collection(kind, program_id):
     plugin.set_content('tvshows')
-    return plugin.finish(view.build_mixed_collection(kind, collection_id, settings))
+    return plugin.finish(view.build_mixed_collection(kind, program_id, settings))
 
 
 @plugin.route('/streams/<program_id>', name='streams')
 def streams(program_id):
-    plugin.set_content('tvshows')
     return plugin.finish(view.build_video_streams(program_id, settings))
 
+@plugin.route('/play_live/<streamUrl>', name='play_live')
+def play_live(streamUrl):
+    return plugin.set_resolved_url({'path': streamUrl})
+
+# Cannot read video new arte tv program API. Blocked by FFMPEG issue #10149
+# @plugin.route('/play_artetv/<program_id>', name='play_artetv')
+# def play_artetv(program_id):
+#     item = api.program_video(settings.language, program_id)
+#     attr = item.get('attributes')
+#     streamUrl=attr.get('streams')[0].get('url')
+#     return plugin.set_resolved_url({'path': streamUrl})
 
 @plugin.route('/play/<kind>/<program_id>', name='play')
 @plugin.route('/play/<kind>/<program_id>/<audio_slot>', name='play_specific')
@@ -113,6 +153,10 @@ def weekly():
     plugin.set_content('tvshows')
     return plugin.finish(view.build_weekly(settings))
 
+@plugin.route('/search', name='search')
+def weekly():
+    plugin.set_content('tvshows')
+    return plugin.finish(view.search(plugin, settings))
 
 # @plugin.route('/broadcast', name='broadcast')
 # def broadcast():