+++ plugin.video.arteplussept/resources/lib/view.py
@@ -1,53 +1,100 @@
+from xbmcswift2 import xbmc
 
 import api
 import mapper
 import hof
 import utils
 
+def build_home_page(plugin, cached_categories, settings):
+    addon_menu = [
+        mapper.create_search_item(),
+        mapper.map_live_video(api.program_video(settings.language, 'LIVE'), settings.quality, '1')]
+    arte_home = api.page(settings.language)
+    for zone in arte_home.get('zones'):
+        menu_item = mapper.map_zone_to_item(zone, cached_categories)
+        if(menu_item):
+            addon_menu.append(menu_item)
+    return addon_menu
 
-def build_categories(settings):
+
+def build_categories(plugin, cached_categories, settings):
     categories = [
+        mapper.create_search_item(),
+        mapper.map_live_video(api.program_video(settings.language, 'LIVE'), settings.quality, '1'),
+        mapper.create_favorites_item(),
+        mapper.create_last_viewed_item(),
         mapper.create_newest_item(),
         mapper.create_most_viewed_item(),
         mapper.create_last_chance_item(),
     ]
-    categories.extend([mapper.map_categories_item(
-        item) for item in api.categories(settings.language)])
+    categories.extend(mapper.map_categories(
+        api.categories(settings.language), settings.show_video_streams, cached_categories))
     # categories.append(mapper.create_creative_item())
     categories.append(mapper.create_magazines_item())
     categories.append(mapper.create_week_item())
-
     return categories
 
 
-def build_category(category_code, settings):
-    category = [mapper.map_category_item(
-        item, category_code) for item in api.category(category_code, settings.language)]
+def build_api_category(category_code, settings):
+    category = [mapper.map_category_item(item, category_code) for item in
+            api.category(category_code, settings.language)]
 
     return category
 
 
+def get_cached_category(category_title, most_viewed_categories):
+    return most_viewed_categories[category_title]
+
+
 def build_magazines(settings):
-    return [mapper.map_generic_item(item, settings.show_video_streams) for item in api.magazines(settings.language)]
+    return [mapper.map_generic_item(item, settings.show_video_streams) for item in
+            api.magazines(settings.language)]
+
+
+def build_favorites(plugin, settings):
+    return [mapper.map_artetv_video(item) for item in
+            api.get_favorites(plugin, settings.language, settings.username, settings.password) or
+            # display an empty list in case of error. error should be display in a notification
+            []]
+
+def add_favorite(plugin, usr, pwd, program_id):
+    if (200 == api.add_favorite(plugin, usr, pwd, program_id)):
+        plugin.notify(msg=plugin.addon.getLocalizedString(30025), image='info')
+    else:
+        plugin.notify(msg=plugin.addon.getLocalizedString(30026), image='error')
+
+def remove_favorite(plugin, usr, pwd, program_id):
+    if (200 == api.remove_favorite(plugin, usr, pwd, program_id)):
+        plugin.notify(msg=plugin.addon.getLocalizedString(30027), image='info')
+    else:
+        plugin.notify(msg=plugin.addon.getLocalizedString(30028), image='error')
+
+
+def build_last_viewed(plugin, settings):
+    return [mapper.map_artetv_video(item) for item in
+            api.last_viewed(plugin, settings.language, settings.username, settings.password) or
+            # display an empty list in case of error. error should be display in a notification
+            []]
 
 
 def build_newest(settings):
-    return [mapper.map_generic_item(item, settings.show_video_streams) for
-            item in api.home_category('mostRecent', settings.language)]
+    return [mapper.map_generic_item(item, settings.show_video_streams) for item in
+            api.home_category('mostRecent', settings.language)]
 
 
 def build_most_viewed(settings):
-    return [mapper.map_generic_item(item, settings.show_video_streams) for
-            item in api.home_category('mostViewed', settings.language)]
+    return [mapper.map_generic_item(item, settings.show_video_streams) for item in
+            api.home_category('mostViewed', settings.language)]
 
 
 def build_last_chance(settings):
-    return [mapper.map_generic_item(item, settings.show_video_streams) for
-            item in api.home_category('lastChance', settings.language)]
+    return [mapper.map_generic_item(item, settings.show_video_streams) for item in
+            api.home_category('lastChance', settings.language)]
 
 
 def build_sub_category_by_code(sub_category_code, settings):
-    return [mapper.map_generic_item(item, settings.show_video_streams) for item in api.subcategory(sub_category_code, settings.language)]
+    return [mapper.map_generic_item(item, settings.show_video_streams) for item in
+            api.subcategory(sub_category_code, settings.language)]
 
 
 def build_sub_category_by_title(category_code, sub_category_title, settings):
@@ -56,11 +103,13 @@ def build_sub_category_by_title(category_code, sub_category_title, settings):
 
     sub_category = hof.find(lambda i: i.get('title') == unquoted_title, category)
 
-    return [mapper.map_generic_item(item, settings.show_video_streams) for item in sub_category.get('teasers')]
+    return [mapper.map_generic_item(item, settings.show_video_streams) for item in
+            sub_category.get('teasers')]
 
 
 def build_mixed_collection(kind, collection_id, settings):
-    return [mapper.map_generic_item(item, settings.show_video_streams) for item in api.collection(kind, collection_id, settings.language)]
+    return [mapper.map_generic_item(item, settings.show_video_streams) for item in
+            api.collection(kind, collection_id, settings.language)]
 
 
 def build_video_streams(program_id, settings):
@@ -76,7 +125,9 @@ def build_video_streams(program_id, settings):
 
 
 def build_stream_url(kind, program_id, audio_slot, settings):
-    return mapper.map_playable(api.streams(kind, program_id, settings.language), settings.quality, audio_slot)
+    return mapper.map_playable(
+        api.streams(kind, program_id, settings.language),
+        settings.quality, audio_slot, mapper.match_hbbtv)
 
 
 _useless_kinds = ['CLIP', 'MANUAL_CLIP', 'TRAILER']
@@ -102,3 +153,19 @@ def build_weekly(settings):
         item, 'info.aired'), reverse=True)
 
     return videos_mapped
+
+def search(plugin, settings):
+    return mapper.map_cached_categories(
+        api.search(settings.language, get_search_query(plugin)))
+
+def get_search_query(plugin):
+    searchStr = ''
+    keyboard = xbmc.Keyboard(searchStr, plugin.addon.getLocalizedString(30012))
+    keyboard.doModal()
+    if (keyboard.isConfirmed() == False):
+        return
+    searchStr = keyboard.getText()
+    if len(searchStr) == 0:
+        return
+    else:
+        return searchStr